I've generated print profiles for the elegoo orangstorm giga to be incorporated with orcaslicer.
I've created machine profiles, print profiles & edited the filament profiles for elegoo to include orangestorm giga.

NOTE: as of now, I do not know how to configure orcaslicer to send different M140 or M190 commands to klipper via the slicer since there are 4 heatbeds on this printer. a typical M140 or M190 command looks like: M140 S60. However klipper will tell orangestorm giga to turn on all heat beds instead of the ones that are actually required. In order to tell klipper to turn on a particular heatbed, the command would need to be sent like this: M140 T0 S60. (this would turn on heatbed 0 or the bottom left heatbed to 60).

I have edited my printer.cfg to automatically take all M140 and m190 commands and determine which heatbed to be turned on based on the min/max x y of the models in orcaslicer.

My printer config file can be located at: https://github.com/TNTBA/OrangeStorm-Giga-Profile-for-Orca/blob/main/klipper_config/printer.cfg